+++ scripts/DataModel.js
@@ -1,0 +1,60 @@
+export const MODEL = {
+  characteristics: {
+    cun: {
+      value: 0
+    },
+    int: {
+      value: 0
+    },
+    per: {
+      value: 0
+    },
+    str: {
+      value: 0
+    },
+    sta: {
+      value: 0
+    },
+    pre: {
+      value: 0
+    },
+    com: {
+      value: 0
+    },
+    dex: {
+      value: 0
+    },
+    qik: {
+      value: 0
+    }
+  }
+};
+// Object.values(CONFIG.ARM5E.character.characteristics).map(e =>  game.i18n.localize(e.label));
+export const CHARACTERISTICS = [
+  "Intelligence",
+  "Perception",
+  "Strength",
+  "Stamina",
+  "Presence",
+  "Communication",
+  "Dexterity",
+  "Quickness"
+];
+// Object.values(CONFIG.ARM5E.magic.arts).map(e =>  e.label);
+export const ARTS = [
+  "Creo",
+  "Intellego",
+  "Muto",
+  "Perdo",
+  "Rego",
+  "Animal",
+  "Aquam",
+  "Auram",
+  "Corpus",
+  "Herbam",
+  "Ignem",
+  "Imaginem",
+  "Mentem",
+  "Terram",
+  "Vim"
+];